wt88047: gps: HAX
Conflicts:
	wt88047/proprietary/vendor/lib/liblocationservice.so

GPS hax is now done at device level.
The qcom location services were removed earlier, to check battery etc. But haxed
liblocationservice.so too for device level gps changes. We keep that file, but
for now we do not use it. If we bring qcom location services back, make sure the
haxed file is used for liblocationservice.so